Changed value of evaluation functions to a float and added a simple evaluation function. Also fixed  syntax error in hashing function and promotion chooser

+++ src/main/java/chess/BoardEvaluator.java
@@ -1,5 +1,5 @@
 package chess;
 
 public interface BoardEvaluator {
-    public int evaluateBoard(Board board);
+    public float evaluateBoard(Board board);
 }
